project2/requirements.txt
pyspark
kafka-python
cassandra-driver
python-dotenv